+++ src/backend/joanie/core/factories.py
@@ -474,3 +474,13 @@ class CertificateFactory(factory.django.DjangoModelFactory):
         Return the order product certificate definition.
         """
         return self.order.product.certificate_definition
+
+
+class CourseWishFactory(factory.django.DjangoModelFactory):
+    """A factory to create a course wish for a user."""
+
+    class Meta:
+        model = models.CourseWish
+
+    course = factory.SubFactory(CourseFactory)
+    owner = factory.SubFactory(UserFactory)